NDEV-1149 Add alternative block_hash fn to cli as well
