+++ evm_loader/program/src/account_storage/backend.rs
@@ -3,23 +3,12 @@ use crate::account_storage::{AccountStorage, ProgramAccountStorage};
 use crate::config::STORAGE_ENTRIES_IN_CONTRACT_ACCOUNT;
 use crate::executor::{OwnedAccountInfo, OwnedAccountInfoPartial};
 use evm::{H160, H256, U256};
-use solana_program::slot_history::Slot;
 use solana_program::{
     pubkey::Pubkey,
     sysvar::slot_hashes::{self, SlotHashes},
 };
 
-fn generate_fake_block_hash(slot: Slot) -> [u8; 32] {
-    let slot_bytes: [u8; 8] = slot.to_be().to_ne_bytes();
-    let non_null_bytes: Vec<_> = slot_bytes.into_iter().skip_while(|&n| n == 0).collect();
-    let non_null_len = non_null_bytes.len();
-    let mut hash = [255; 32];
-    hash[32 - 1 - non_null_len] = 0;
-    for i in 0..non_null_len {
-        hash[32 - non_null_len + i] = non_null_bytes[i];
-    }
-    hash
-}
+use super::generate_fake_block_hash;
 
 impl<'a> AccountStorage for ProgramAccountStorage<'a> {
     fn neon_token_mint(&self) -> &Pubkey {
